remove rest api handlers since moved to grpc
